+++ DawnLib/src/Utils/MiscScripts/ExtraEvents/ExtraScanEvents.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+using UnityEngine.Events;
+
+namespace Dawn.Utils;
+[RequireComponent(typeof(ScanNodeProperties))]
+public class ExtraScanEvents : MonoBehaviour
+{
+    [SerializeField]
+    internal UnityEvent _onScan;
+}