+++ Thunderstore/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.1.6
 - Improved error handling with corrupted save data and tagging.
 - Fixed an issue with applying item spawn group tags
+- Added `ExtraScanEvents` with an `On Scan` event (requires `Scan Node Properties`)
 
 ## v0.1.5